add IReadOnlyList<T> support to most database types

+++ libloc/V1/DatabaseV1.cs
@@ -77,7 +77,7 @@ namespace libloc.V1
         public IAddressLocatedNetwork ResolveAddress(IPAddress address)
         {
             var depth = -1;
-            uint nextNodeIndex = 0;
+            int nextNodeIndex = 0;
 
             var mappedAddress = address.MapToIPv6().GetAddressBytes();
             Span<byte> networkAddress = stackalloc byte[mappedAddress.Length];
@@ -97,7 +97,7 @@ namespace libloc.V1
                 AddressUtils.SetAddressBit(networkAddress, depth, bit);
 
                 node = _networkTree.ElementAt(nextNodeIndex);
-                nextNodeIndex = BinaryUtils.EnsureEndianness(bit == 0 ? node.zero : node.one);
+                nextNodeIndex = (int)BinaryUtils.EnsureEndianness(bit == 0 ? node.zero : node.one);
             } while (nextNodeIndex > 0);
 
             if (!node.IsLeaf)
